+++ Restaurants.Application/CQRS/Restaurants/RestaurantsCommands/DeleteRestaurantCommand/DeleteRestaurantCommandHandler.cs
@@ -1,5 +1,6 @@
 ﻿using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
 using Restaurants.Domain.Repositories;
@@ -7,7 +8,8 @@ using Restaurants.Domain.Repositories;
 namespace Restaurants.Application.CQRS.Restaurants.RestaurantsCommands.DeleteRestaurantCommand;
 
 public class DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommandHandler> logger,
-    IRestaurantRepository repository) : IRequestHandler<Restaurants.RestaurantsCommands.DeleteRestaurantCommand.DeleteRestaurantCommand>
+    IRestaurantRepository repository,
+    IRestaurantAuthorizationService restaurantAuthorizationService) : IRequestHandler<Restaurants.RestaurantsCommands.DeleteRestaurantCommand.DeleteRestaurantCommand>
 {
     public async Task Handle(Restaurants.RestaurantsCommands.DeleteRestaurantCommand.DeleteRestaurantCommand request, CancellationToken cancellationToken)
     {
@@ -16,6 +18,11 @@ public class DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommandHandl
         if (restaurant is null)
         {
             throw new NotFoundException(nameof(Restaurant), request.Id.ToString());
+        }
+
+        if (!restaurantAuthorizationService.Authorize(restaurant, ResourceOperations.Delete))
+        {
+            throw new ForbidException();
         }
 
         await repository.DeleteAsync(restaurant);
